web_app/requirements.txt
Flask==2.0.1
psycopg2
psycopg2-binary
flask-sqlalchemy